添加Service层

+++ OctopusParent/ControlCenter/src/test/java/com/octopus/eureka/control/ControlCenterApplicationTests.java
@@ -10,6 +10,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 import javax.annotation.Resource;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
@@ -25,7 +26,7 @@ public class ControlCenterApplicationTests {
             BuyBo test = new BuyBo();
             test.setBusinessCode("110");
             test.setCustomerId("1111111");
-            test.setOrderSeq(100);
+            test.setOrderSeq(BigInteger.valueOf(100));
             test.setProductId("01000");
             test.setTransactionAmount(new BigDecimal(100.00));
             test.setDestiny("order");
